backend/requirements.txt
Django==2.2
djangorestframework==3.9.2
django-cors-headers==2.5.2
